Remove warning suppression.
This was added in https://github.com/google/ExoPlayer/commit/f9f228994899adba712751af5aa4c7e4d1a6ee5f as part of a LSC.

The RequiresNonNull annotation doesn't work anymore (it doesn't
recognize the outer class member and instead tries to find
the same variable on the inner class). So instead of suppressing
the warning of the non-fulfilled precondition, we can just
check the non-nullness directly and remove the precondition.

PiperOrigin-RevId: 367593941