Rollback of https://github.com/google/ExoPlayer/commit/a36ac0108e3ecba3b0f24b04e472f265f386bce6
*** Original commit ***

Rollback of https://github.com/google/ExoPlayer/commit/c1deeb65f01e79a833564520347d76e7073844db

*** Original commit ***

Prevent creation of new sessions if the Timeline is empty.

We currently create sessions based on the placeholder window
index. This shouldn't be needed as we now set a non-empty
timeline as soon as the first MediaItem...

***

PiperOrigin-RevId: 369609523