Update javadoc for 2.14.2
This also deletes several files that should have been deleted as part
of the 2.14.0 release - both javadoc for classes that were removed in
that release and all the *-frame.html files that no longer seem
to be produced by the javadoc compiler. Also a jquery file
(jquery-1.10.2.js) that's no longer produced by the compiler.

#minor-release

PiperOrigin-RevId: 385997294