Use JUnit's Assume to skip Widevine GTS tests on non-GMS devices
DashTestRunner.setWidevineInfo instantiates a Widevine MediaDrm to check
for L1 support, which fails on GMS devices that don't support Widevine.

The Widevine tests should be skipped on these devices, but the
setWidevineInfo call happens in @Before before the skipping logic
inside the test.

This change moves the skipping logic to the beginning of @Before, and
uses Assume to ensure the test isn't run.

This will also report the test as 'skipped due to unmet assumption'
which is more precise than 'passed'.

PiperOrigin-RevId: 399452734