Timestamp init wait occurs after dataSource.open()
Opening the `DataSource` is one of the longer operations in the `Loader` sequence, as it requires a round trip to the origin server.   This change allows all the potential `Loader` threads to perform this operation before one of them is forced to wait on shared TimestampAdjuster initialization.

Also, the initialization segment load will never produce media samples, so there is no need for it to wait for `TimestampAdjuster` initialization.